# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Voxlab public speaking educational platform",
        "title": "Voxlab API",
        "termsOfService": "http://voxlab.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@voxlab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all registered users. Admin only. Includes profile data (name, email, XP, streak).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "List all users (admin)",
                "responses": {
                    "200": {
                        "description": "Lista de todos los usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.ListUsersResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "403": {
                        "description": "Solo administradores pueden listar usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.ForbiddenError"
                        }
                    },
                    "500": {
                        "description": "Error al obtener los usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single user's profile data. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.GetUserResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "403": {
                        "description": "Solo administradores pueden ver usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.ForbiddenError"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modifies a user's role, name, email, or other fields. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos en la solicitud",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "403": {
                        "description": "Solo administradores pueden modificar usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.ForbiddenError"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently removes a user account. Admin only. ⚠️ This cannot be undone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.DeleteUserResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "403": {
                        "description": "Solo administradores pueden eliminar usuarios",
                        "schema": {
                            "$ref": "#/definitions/resources.ForbiddenError"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al eliminar el usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resources.BadRequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Solicitud inválida — verifique los campos enviados"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.DeleteUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "resources.ForbiddenError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "No tiene permisos para realizar esta acción"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.GetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/services.AdminUserData"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "resources.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.ListUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.AdminUserData"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.NotFoundError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Recurso no encontrado"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.UnauthorizedError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Token de autenticación no proporcionado o inválido"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/services.AdminUserData"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "services.AdminUserData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-06-01T12:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "lives": {
                    "type": "integer",
                    "example": 5
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "streak_days": {
                    "type": "integer",
                    "example": 3
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-18T12:00:00Z"
                },
                "xp": {
                    "type": "integer",
                    "example": 850
                }
            }
        },
...
        "services.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}